Test that sxe-ssl works with multiple buffers.
Call SSL_write() until we run out of buffers to send, or it returns an error
(likely that it wants to write). On the Mac, it seems to take a lot of data
before it fails, so buffers are a convenient way to hammer the same data out
the network until it finally gives in.

This improves coverage slightly: the SSL_ERROR_WANT_WRITE case is now covered.