Do not use sxe_mmap_open() on argv[0].
Using sxe_mmap_open() on argv[0] works on Mac OS X, but doesn't work on Linux.
sxe_mmap_open() tries to open the file O_RDWR, which fails with EBUSY because
the file is in use (it's the one trying to open itself).

Instead, just open ourselves and read into an allocated buffer.